scripts: add StarFive JH7110 STGCRG definitions
Adds a script for StarFive JH7110 STGCRG clock register definitions.
